+++ code/glo_loader.h
@@ -34,11 +34,11 @@
 class GloLoader : public TileLoader {
 public:
   // minLat and minLng name the SW corner of the tile, in degrees
-  virtual Tile *loadTile(const std::string &directory, float minLat, float minLng);
+  virtual Tile *loadTile(const std::string &directory, double minLat, double minLng);
 
 private:
   // Width in samples for a tile at the given latitude
-  int getWidthForLatitude(float minLat) const;
+  int getWidthForLatitude(double minLat) const;
 };
 
 #endif  // _GLO_LOADER_H_
